Minimize webhook ledger and logging
